+++ R/DBResult.R
@@ -356,7 +356,6 @@ setMethod("dbGetInfo", "DBIResult", function(dbObj, ...) {
 #' nrow(dbReadTable(con, "iris"))
 #'
 #' dbDisconnect(con)
-setGeneric(
-  "dbBind",
-  function(res, params, ...) standardGeneric("dbBind")
+setGeneric("dbBind",
+  def = function(res, params, ...) standardGeneric("dbBind")
 )