+++ R/deprecated.R
@@ -53,11 +53,9 @@
 #'
 #'   `isSQLKeyword` returns a logical vector parallel to `name`.
 #' @export
-setGeneric(
-  "make.db.names",
+setGeneric("make.db.names",
+  def = function(dbObj, snames, keywords = .SQL92Keywords, unique = TRUE, allow.keywords = TRUE, ...) standardGeneric("make.db.names"),
   signature = c("dbObj", "snames"),
-  function(dbObj, snames, keywords = .SQL92Keywords, unique = TRUE, allow.keywords = TRUE, ...)
-    standardGeneric("make.db.names"),
   valueClass = "character"
 )
 
@@ -102,11 +100,10 @@ make.db.names.default <- function(snames, keywords = .SQL92Keywords,
 
 #' @rdname make.db.names
 #' @export
-setGeneric(
-  "isSQLKeyword",
-  signature = c("dbObj", "name"),
-  function(dbObj, name, keywords = .SQL92Keywords, case = c("lower", "upper", "any")[3], ...)
+setGeneric("isSQLKeyword",
+  def = function(dbObj, name, keywords = .SQL92Keywords, case = c("lower", "upper", "any")[3], ...)
     standardGeneric("isSQLKeyword"),
+  signature = c("dbObj", "name"),
   valueClass = "logical"
 )
 
@@ -137,9 +134,8 @@ isSQLKeyword.default <- function(name, keywords = .SQL92Keywords,
 ## (See pp. 22 and 23 in X/Open SQL and RDA, 1994, isbn 1-872630-68-8)
 
 #' @export
-setGeneric(
-  "SQLKeywords",
-  function(dbObj, ...) standardGeneric("SQLKeywords"),
+setGeneric("SQLKeywords",
+  def = function(dbObj, ...) standardGeneric("SQLKeywords"),
   valueClass = "character"
 )
 
@@ -200,8 +196,7 @@ setMethod("SQLKeywords", "missing",
 #' @inheritParams dbGetQuery
 #' @keywords internal
 #' @export
-setGeneric(
-  "dbCallProc",
+setGeneric("dbCallProc",
   def = function(conn, ...) {
     .Deprecated()
     standardGeneric("dbCallProc")
@@ -232,8 +227,7 @@ dbGetDBIVersion <- function() {
 #' @keywords internal
 #' @param flds a field description object as returned by `dbColumnInfo`.
 #' @export
-setGeneric(
-  "dbSetDataMappings",
+setGeneric("dbSetDataMappings",
   def = function(res, flds, ...) {
     .Deprecated()
     standardGeneric("dbSetDataMappings")